create cache.ts
