+++ src/use-swr.ts
@@ -26,10 +26,11 @@ import defaultConfig, {
   CONCURRENT_PROMISES_TS,
   FOCUS_REVALIDATORS,
   CACHE_REVALIDATORS,
-  MUTATION_TS,
-  cacheGet,
-  cacheSet
+  MUTATION_TS
 } from './config'
+
+import { cacheGet, cacheSet } from './cache'
+
 import SWRConfigContext from './swr-config-context'
 import isDocumentVisible from './libs/is-document-visible'
 import useHydration from './libs/use-hydration'
